invoice_system/requirements.txt
Django==4.2
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework==3.14.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
reportlab==4.0.9
matplotlib==3.8.2
urllib3==2.1.0
pathlib==1.0.1
Pillow==10.1.0
pytz==2023.3.post1